Added Student model.
Moved all universal attributes to Identity

After further discussion, we decided to move all attributes that could
be associated to any type of user to the identity model. In doing so,
changes to the spec file were necessary. At this early stage of
development the database was dropped and the migration files were
modified rather than adding unnecessary additional migrations.